config.go
package config

import (
	"encoding/json"
	"io/ioutil"

	"github.com/Jeffail/gabs"
)

// Value config item value
type Value interface{}

// Config golang json config object
type Config struct {
	container *gabs.Container
}

// New load config from json bytes
func New(source []byte) (*Config, error) {
	parsed, err := gabs.ParseJSON(source)

	if err != nil {
		return nil, err
	}

	return &Config{
		container: parsed,
	}, nil
}

// Reload reload config from source bytes
func (config *Config) Reload(source []byte) error {
	parsed, err := gabs.ParseJSON(source)

	if err != nil {
		return err
	}

	config.container = parsed

	return nil
}

// Get get config value
func (config *Config) Get(path string) Value {
	return config.container.Path(path).Data()
}

// GetObject get config value as object
func (config *Config) GetObject(path string, v interface{}) error {
	bytes, err := json.Marshal(config.container.Path(path).Data())

	if err != nil {
		return err
	}

	return json.Unmarshal(bytes, v)
}

// GetInt64 get config value as Int
func (config *Config) GetInt64(path string, defaultval int64) int64 {
	value, ok := config.container.Path(path).Data().(float64)

	if !ok {
		return defaultval
	}

	return int64(value)
}

// GetString get config value as Int
func (config *Config) GetString(path string, defaultval string) string {
	value, ok := config.container.Path(path).Data().(string)

	if !ok {
		return defaultval
	}

	return value
}

// Has check if has config item indicate by path
func (config *Config) Has(path string) bool {
	return config.container.ExistsP(path)
}

var config = &Config{}

// Load load config from source bytes
func Load(source []byte) {
	config.Reload(source)
}

// LoadFromFile load config from config json file
func LoadFromFile(filepath string) error {
	data, err := ioutil.ReadFile(filepath)

	if err != nil {
		return err
	}

	config.Reload(data)

	return nil
}

// Get global method, get config value from global config object
func Get(path string) Value {
	return config.Get(path)
}

// Has global method, check if global config object has the config item
func Has(path string) bool {
	return config.Has(path)
}

// GetInt64 get config value as Int
func GetInt64(path string, defaultval int64) int64 {
	return config.GetInt64(path, defaultval)
}

// GetString get config value as String
func GetString(path string, defaultval string) string {
	return config.GetString(path, defaultval)
}

// GetObject get config value as object
func GetObject(path string, v interface{}) error {
	return config.GetObject(path, v)
}
